% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{aFiles}
\alias{aFiles}
\alias{aFiles<-}
\alias{aFiles,ISA-method}
\alias{aFiles<-,ISA-method}
\title{Get and set aFiles.}
\usage{
aFiles(x)

aFiles(x) <- value

\S4method{aFiles}{ISA}(x)

\S4method{aFiles}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A list of data.frames containing the assay files.
In each data.frame at least the following columns are present:\if{html}{\out{<div class=".">}}\preformatted{}\if{html}{\out{</div>}}}
}
\value{
A list of data.frames containing the assay files.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the list of aFiles data.frame in an object of
\code{\link{ISA-class}}.
}
