% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{assayTab-class}
\alias{assayTab-class}
\alias{assayTab}
\alias{msAssayTab-class}
\alias{msAssayTab}
\alias{microarrayAssayTab-class}
\alias{microarrayAssayTab}
\title{S4 Class assayTab, initialization methods}
\description{
An S4 class to store information from an assay files.
}
\section{Slots}{

\describe{
\item{\code{path}}{A length-one character vector containing the path to the ISA-Tab
data set.}

\item{\code{sFileName}}{A length-one character vector containing the study
file name (by definition starting with \strong{s_} and ending at \strong{.txt}).}

\item{\code{sIdentifier}}{A length-one character vector containing the study
identifier.}

\item{\code{aFileName}}{A length-one character vector containing the assay
file name (by definition starting with \strong{a_} and ending at \strong{.txt}).}

\item{\code{aFile}}{A data.frame containing the contents of the Assay Table file.}

\item{\code{aTechType}}{A length-one character vector containing the assay
technology type.}

\item{\code{aMeasType}}{A length-one character vector containing the assay
measurement type.}
}}

\keyword{classes}
