% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-getinfo.R
\name{getAssayInfo}
\alias{getAssayInfo}
\title{Retrieve the Assay File Name(s) per Study from an ISA object.}
\usage{
getAssayInfo(isaObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}
}
\value{
A named character vector of a list containing the Assay File Name(s)
        for each Study Indentifier. The name of the character vector or names
        of the list elements represent(s) the Study Identifier(s).
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the Assay File Name(s)
linked to the Study Identifier(s) per Study.
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFileNames}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayMeasurementTypesPerStudy}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTabs}()},
\code{\link{getAssayTechnologyTypesPerStudy}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamplesPerStudy}()},
\code{\link{getSamples}()},
\code{\link{getStudyInfo}()},
\code{\link{getTreatmentGroups}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
