% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-getinfo.R
\name{getAssayMeasurementTypesPerStudy}
\alias{getAssayMeasurementTypesPerStudy}
\title{Retrieve the Study Assay Measurement Types per Study from an ISA object.}
\usage{
getAssayMeasurementTypesPerStudy(isaObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}
}
\value{
A list of Study Identifiers. Each element of the list, named by a
        Study Identifier, contains a character vector specifying the Study
        Assay Measurement Types belonging to a specific study (marked by its
        Study Identifier).
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the Study Assay
Measurement Types per Study.
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFileNames}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayInfo}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTechnologyTypesPerStudy}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamplesPerStudy}()},
\code{\link{getSamples}()},
\code{\link{getStudyInfo}()},
\code{\link{getTreatmentGroups}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
