% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-getinfo.R
\name{getAssayTechnologyTypesPerStudy}
\alias{getAssayTechnologyTypesPerStudy}
\title{Retrieve the Study Assay Technology Types per Study from an ISA object.}
\usage{
getAssayTechnologyTypesPerStudy(isaObject)
}
\value{
A list of Study Identifiers. Each element of the list contains a
        character vector specifying the Study Assay Technology Types
        belonging to that specific study (marked by its Study Identifier).
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the Study Assay
Technology Types per Study.
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFileNames}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayInfo}()},
\code{\link{getAssayMeasurementTypesPerStudy}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTabs}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamplesPerStudy}()},
\code{\link{getSamples}()},
\code{\link{getStudyInfo}()},
\code{\link{getTreatmentGroups}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
