% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-getinfo.R
\name{getSamplesPerStudy}
\alias{getSamplesPerStudy}
\title{Retrieve the unique Sample Names per Study from an ISA object.}
\usage{
getSamplesPerStudy(isaObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}
}
\value{
A list of Study Identifiers. Each element of the list, named by the
Study Identifier, contains a character vector specifying the unique
Sample Names belonging to that specific study.
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the unique Sample
Names per Study.
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFileNames}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayInfo}()},
\code{\link{getAssayMeasurementTypesPerStudy}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTabs}()},
\code{\link{getAssayTechnologyTypesPerStudy}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamples}()},
\code{\link{getStudyInfo}()},
\code{\link{getTreatmentGroups}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
