% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-getinfo.R
\name{getStudyInfo}
\alias{getStudyInfo}
\title{Retrieve the Study Indentifier(s) and Study File Name(s) from an ISA object.}
\usage{
getStudyInfo(isaObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}
}
\value{
A named character vector containing the Study File Name(s) and the
        name(s) representing the Study Identifier(s).
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the Study
Identifier(s) and Study File Name(s) as contained in the Investigation.
To directly access the Study Identifier(s) use the names() function, e.g.
names(getStudyInfo(isaObject)).
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFileNames}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayInfo}()},
\code{\link{getAssayMeasurementTypesPerStudy}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTechnologyTypesPerStudy}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamplesPerStudy}()},
\code{\link{getSamples}()},
\code{\link{getTreatmentGroups}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
