% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo.R
\name{getTreatmentGroups}
\alias{getTreatmentGroups}
\title{Retrieve Sample Names for each treatment per Study File from an ISA object.}
\usage{
getTreatmentGroups(isaObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}
}
\value{
A list of lists, where each list element, named by the Study
Identifier, contains a list of treatments. Each element in the list
of treatments contains the Sample Names used in that specific Study
using a specific treatment, i.e. combination of Factor Values.
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the Sample Names for
each treatment, as unique combination of specific Factor Values, per Study
File. The treaments can be retrieved by the \code{\link{getTreatments}}
function.
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFileNames}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayInfo}()},
\code{\link{getAssayMeasurementTypesPerStudy}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTabs}()},
\code{\link{getAssayTechnologyTypesPerStudy}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamplesPerStudy}()},
\code{\link{getSamples}()},
\code{\link{getStudyInfo}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
