% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iContacts}
\alias{iContacts}
\alias{iContacts<-}
\alias{iContacts,ISA-method}
\title{Get and set iContacts.}
\usage{
iContacts(x)

iContacts(x) <- value

\S4method{iContacts}{ISA}(x)
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A data.frame containing the investigation contacts
information. A data.frame with the columns
\verb{Investigation Person Last Name}, \verb{Investigation Person First Name}, \verb{Investigation Person Mid Initials}, \verb{Investigation Person Email}, \verb{Investigation Person Phone}, \verb{Investigation Person Fax}, \verb{Investigation Person Address}, \verb{Investigation Person Affiliation}, \verb{Investigation Person Roles}, \verb{Investigation Person Roles Term Accession Number}, \verb{Investigation Person Roles Term Source REF}.}
}
\value{
A data.frame containing the investigation contacts information.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the iContacts data.frame in an object of
\code{\link{ISA-class}}.
}
