% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iFileName}
\alias{iFileName}
\alias{iFileName<-}
\alias{iFileName,ISA-method}
\title{Get and set the file name for the investigation file in an object of
\code{\link{ISA-class}}.}
\usage{
iFileName(x)

iFileName(x) <- value

\S4method{iFileName}{ISA}(x)
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A length-one character vector indicating the name of the
investigation file, a string starting with "i_" and ending in ".txt"}
}
\value{
The file name of the investigation file.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the file name for the investigation file in an object of
\code{\link{ISA-class}}.
}
