% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{invest}
\alias{invest}
\alias{invest<-}
\alias{invest,ISA-method}
\title{Get and set invest.}
\usage{
invest(x)

invest(x) <- value

\S4method{invest}{ISA}(x)
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A data.frame containing the investigation information. A
data.frame in which at least the following columns are present:
\verb{Investigation Identifier}, \verb{Investigation Title}, \verb{Investigation Description}, \verb{Investigation Submission Date}, \verb{Investigation Public Release Date}.}
}
\value{
A data.frame containing the investigation information.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the investigation data.frame in an object of
\code{\link{ISA-class}}.
}
