% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iPubs}
\alias{iPubs}
\alias{iPubs<-}
\alias{iPubs,ISA-method}
\title{Get and set iPubs.}
\usage{
iPubs(x)

iPubs(x) <- value

\S4method{iPubs}{ISA}(x)
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A data.frame containing the investigation publications
information. data.frame in which at least the following columns are present:
\verb{Investigation PubMed ID}, \verb{Investigation Publication DOI}, \verb{Investigation Publication Author List}, \verb{Investigation Publication Title}, \verb{Investigation Publication Status}, \verb{Investigation Publication Status Term Accession Number}, \verb{Investigation Publication Status Term Source REF}.}
}
\value{
A data.frame containing the investigation publications information.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the iPubs data.frame in an object of
\code{\link{ISA-class}}.
}
