% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R
\docType{class}
\name{ISA-class}
\alias{ISA-class}
\alias{ISA}
\alias{initialize,ISA-method}
\alias{ISA-initialize}
\alias{[,ISA,character-method}
\alias{extractandreplace-methods}
\alias{[<-,ISA,character-method}
\title{S4 Class ISA, initialization, extract- and replace-methods}
\usage{
\S4method{initialize}{ISA}(.Object, path)

\S4method{[}{ISA,character}(x, i)

\S4method{[}{ISA,character}(x, i) <- value
}
\arguments{
\item{.Object}{character, name of the object of class \linkS4class{ISA} to
be initialized}

\item{path}{length-one character vector containing the path to the ISA-Tab
files of the dataset.}

\item{x}{S4 object of class \linkS4class{ISA}.}

\item{i}{character, name of a slot to extract or replace.}

\item{value}{ANY, replacement value for the slotname specified by \code{i}.}
}
\description{
An S4 class to store information from an ISA-Tab dataset, including an
investigation file, one or more study files, and one or more assay files for
each study file.

When creating a new object of class \linkS4class{ISA} via:
\code{object <- new(Class = "ISA", path)}, the function
\code{initialize(.Object, path)} is called to initialize and create the
actual object. The \code{initialize-method} is seldomly used as a function
itself.
}
\details{
\code{x[i]} defines a method to extract the information in the slot of an S4
object of class \\linkS4class{ISA} by using a character vector specifying the
name of the slot. Via \code{x[i] <- value} the value(s) in a slot of an S4
class object of class \linkS4class{ISA} can be replaced by means of
specifying the name of the slot as a character vector. Below in the slot
descriptions the character vectors to access specific slots are given.
}
\section{Slots}{

\describe{
\item{\code{path}}{A length-one character vector containing the path to the ISA-Tab
dataset. The character vector to extract or replace this slot is
"path".}

\item{\code{iFileName}}{A length-one character vector containing the investigation
filename (by definition starting with \strong{\emph{i_}} and
ending at \strong{\emph{.txt}}). The character vector to
extract or replace this slot is "Investigation File Name".}

\item{\code{oSR}}{A data.frame containing the "ONTOLOGY SOURCE REFERENCE" section of
the investigation file. The character vector to extract or replace
this slot is "ONTOLOGY SOURCE REFERENCE".}

\item{\code{invest}}{A data.frame containing the "INVESTIGATION" section of
the investigation file. The character vector to extract or
replace this slot is "INVESTIGATION".}

\item{\code{iPubs}}{A data.frame containing the "INVESTIGATION PUBLICATIONS" section
of the investigation file. The character vector to extract or
replace this slot is "INVESTIGATION PUBLICATIONS".}

\item{\code{iContacts}}{A data.frame containing the "INVESTIGATION CONTACTS" section
of the investigation file. The character vector to extract or
replace this slot is "INVESTIGATION CONTACTS".}

\item{\code{study}}{A list of data.frames containing the "STUDY" sections of the
investigation file. Each study has its own section and the study
identifier is used to name each element in the list. The character
vector to extract or replace this slot is "STUDY".}

\item{\code{sDD}}{A list of data.frames containing the "STUDY DESIGN DESCRIPTORS"
sections of the investigation file. Each study has its own section
and the study identifier is used to name each element in the list.
The character vector to extract or replace this slot is "STUDY
DESIGN DESCRIPTORS".}

\item{\code{sPubs}}{A list of data.frames containing the "STUDY PUBLICATIONS"
sections of the investigation file. Each study has its own
section and the study identifier is used to name each element in
the list. The character vector to extract or replace this slot is
"STUDY PUBLICATIONS".}

\item{\code{sFacts}}{A list of data.frames containing the "STUDY FACTORS" sections of
the investigation file. Each study has its own section and the
study identifier is used to name each element in the list. The
character vector to extract or replace this slot is "STUDY
FACTORS".}

\item{\code{sAssays}}{A list of data.frames containing the "STUDY ASSAYS" sections of
the investigation file. Each study has its own section and the
study identifier is used to name each element in the list. The
character vector to extract or replace this slot is "STUDY
ASSAYS".}

\item{\code{sProts}}{A list of data.frames containing the "STUDY PROTOCOLS" sections
of the investigation file. Each study has its own section and
the study identifier is used to name each element in the list.
The character vector to extract or replace this slot is "STUDY
PROTOCOLS".}

\item{\code{sContacts}}{A list of data.frames containing the "STUDY CONTACTS"
sections of the investigation file. Each study has its own
section and the study identifier is used to name each element
in the list. The character vector to extract or replace this
slot is "STUDY CONTACTS".}

\item{\code{sFiles}}{A list of data.frames containing the "Study Files", containing
the contents of the Study Table files belonging to the studies
in the investigation. Each study has one Study Table file. The
character vector to extract or replace this slot is "Study
Files".}

\item{\code{aFiles}}{A list of data.frames containing the "Assay Files", containing
the contents of the Assay Table files belonging to the studies
in the investigation. Each study can have multiple Assay Table
files. Each element of the list is named by the Assay File Name.
The character vector to extract or replace this slot is "Assay
Files".}
}}

\author{
Maikel Verouden
}
\keyword{classes}
