% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{isaPath}
\alias{isaPath}
\alias{isaPath<-}
\alias{isaPath,ISA-method}
\alias{isaPath<-,ISA-method}
\title{Get and set path}
\usage{
isaPath(x)

isaPath(x) <- value

\S4method{isaPath}{ISA}(x)

\S4method{isaPath}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A length-one character vector indicating the isaPath to an
accessible directory on the system.}
}
\value{
The path to the folder for the object of class \linkS4class{ISA}.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the file path for an object of class \linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract path.
isaPath(isaObject1)

}
