% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{path}
\alias{path}
\alias{path<-}
\alias{path,ISA-method}
\alias{path<-,ISA-method}
\title{Get and set path}
\usage{
path(x)

path(x) <- value

\S4method{path}{ISA}(x)

\S4method{path}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A length-one character vector indicating the path to an
accessible directory on the system.}
}
\value{
The path to the folder for the object of class \linkS4class{ISA}.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the file path for an object of class \linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract path.
path(isaObject1)

}
