% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssay.R
\docType{methods}
\name{processAssay}
\alias{processAssay}
\alias{processAssay,ISA,assayTab-method}
\alias{processAssay,ISA,msAssayTab-method}
\title{Process assay tab data}
\usage{
processAssay(isaObject, aTabObject)

\S4method{processAssay}{ISA,assayTab}(isaObject, aTabObject)

\S4method{processAssay}{ISA,msAssayTab}(isaObject, aTabObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}

\item{aTabObject}{An object of the \code{\link{msAssayTab-class}}.}
}
\description{
Process data from assay tab files

Process data from assay tab files with technology type mass spectrometry
(ms). Processing those files requires the xcms package to be installed.
}
