% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sAssays}
\alias{sAssays}
\alias{sAssays<-}
\alias{sAssays,ISA-method}
\alias{sAssays<-,ISA-method}
\title{Get and set sAssays.}
\usage{
sAssays(x)

sAssays(x) <- value

\S4method{sAssays}{ISA}(x)

\S4method{sAssays}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A list of data.frames containing the study assays.
In each data.frame at least the following columns are present:
\verb{Study Assay Measurement Type}, \verb{Study Assay Measurement Type Term Accession Number}, \verb{Study Assay Measurement Type Term Source REF}, \verb{Study Assay Technology Type}, \verb{Study Assay Technology Type Term Accession Number}, \verb{Study Assay Technology Type Term Source REF}, \verb{Study Assay Technology Platform}, \verb{Study Assay File Name}.}
}
\value{
A list of data.frames containing the study assays.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the list of sAssays data.frame in an object of
\code{\link{ISA-class}}.
}
