% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sContacts}
\alias{sContacts}
\alias{sContacts<-}
\alias{sContacts,ISA-method}
\alias{sContacts<-,ISA-method}
\title{Get and set sContacts.}
\usage{
sContacts(x)

sContacts(x) <- value

\S4method{sContacts}{ISA}(x)

\S4method{sContacts}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study contacts.
In each data.frame at least the following columns are present:
'Study Person Last Name', 'Study Person First Name', 'Study Person Mid Initials', 'Study Person Email', 'Study Person Phone', 'Study Person Fax', 'Study Person Address', 'Study Person Affiliation', 'Study Person Roles', 'Study Person Roles Term Accession Number', 'Study Person Roles Term Source REF'.}
}
\value{
A list of data.frames containing the study contacts.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of sContacts data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of sContacts data.frames.
sContacts(isaObject1)

}
