% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sDD}
\alias{sDD}
\alias{sDD<-}
\alias{sDD,ISA-method}
\alias{sDD<-,ISA-method}
\title{Get and set sDD.}
\usage{
sDD(x)

sDD(x) <- value

\S4method{sDD}{ISA}(x)

\S4method{sDD}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A list of data.frames containing the study design descriptors.
In each data.frame at least the following columns are present:
\verb{Study Design Type}, \verb{Study Design Type Term Accession Number}, \verb{Study Design Type Term Source REF}.}
}
\value{
A list of data.frames containing the study design descriptors.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the list of sDD data.frame in an object of
\code{\link{ISA-class}}.
}
