% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sFacts}
\alias{sFacts}
\alias{sFacts<-}
\alias{sFacts,ISA-method}
\alias{sFacts<-,ISA-method}
\title{Get and set sFacts.}
\usage{
sFacts(x)

sFacts(x) <- value

\S4method{sFacts}{ISA}(x)

\S4method{sFacts}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A list of data.frames containing the study factors.
In each data.frame at least the following columns are present:
\verb{Study Factor Name}, \verb{Study Factor Type}, \verb{Study Factor Type Term Accession Number}, \verb{Study Factor Type Term Source REF}.}
}
\value{
A list of data.frames containing the study factors.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the list of sFacts data.frame in an object of
\code{\link{ISA-class}}.
}
