% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sProts}
\alias{sProts}
\alias{sProts<-}
\alias{sProts,ISA-method}
\alias{sProts<-,ISA-method}
\title{Get and set sProts.}
\usage{
sProts(x)

sProts(x) <- value

\S4method{sProts}{ISA}(x)

\S4method{sProts}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study protocols.
In each data.frame at least the following columns are present:
\verb{Study Protocol Name}, \verb{Study Protocol Type}, \verb{Study Protocol Type Term Accession Number}, \verb{Study Protocol Type Term Source REF}, \verb{Study Protocol Description}, \verb{Study Protocol URI}, \verb{Study Protocol Version}, \verb{Study Protocol Parameters Name}, \verb{Study Protocol Parameters Name Term Accession Number}, \verb{Study Protocol Parameters Name Term Source REF}, \verb{Study Protocol Components Name}, \verb{Study Protocol Components Type}, \verb{Study Protocol Components Type Term Accession Number}, \verb{Study Protocol Components Type Term Source REF}.}
}
\value{
A list of data.frames containing the study protocols.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of sProts data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of sProts data.frames.
sProts(isaObject1)

}
