% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-validityFunctions.R
\name{validISAObject}
\alias{validISAObject}
\title{Check the validity of an object of class ISA.}
\usage{
validISAObject(object)
}
\arguments{
\item{object}{An object of class \linkS4class{ISA}.}
}
\value{
TRUE or an error message.
}
\description{
The \code{validISAObject} function checks whether an object of class
\linkS4class{ISA} is a valid object (in this case TRUE is returned, otherwise
an error message will appear). An object of the \code{\link{ISA-class}} is
valid when:
\itemize{
 \item There is only one investigation file, which name starts with
       \strong{\emph{i_}} and ends with \strong{\emph{.txt}}, present in the
       folder containing the ISA-Tab files.
}
}
\examples{
## Example BII data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/BII-I-1",
                                          package = "isatabr")))
validISAObject(isaObject1)

}
\seealso{
\linkS4class{ISA}
}
