% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeAssayFiles}
\alias{writeAssayFiles}
\title{Write assay files.}
\usage{
writeAssayFiles(
  isaObject,
  assayFilenames = getAssayFileNames(isaObject),
  path = getwd()
)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}

\item{assayFilenames}{A character vector indicating the study files that
should be written. Default all study files in isaObject
are written.}

\item{path}{A character vector with the name of the directory to which the
file(s) should be written. The default value is the current
working directory.}
}
\description{
Write assay files.
}
