% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeInvestigationFile}
\alias{writeInvestigationFile}
\title{Write investigation file.}
\usage{
writeInvestigationFile(isaObject, path = getwd())
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}

\item{path}{A character vector with the name of the directory to which the
ISA-Tab investigation file should be written. The default
value is the current working directory.}
}
\description{
Write investigation file.
}
