% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeISAtab}
\alias{writeISAtab}
\title{Write ISA-Tab object.}
\usage{
writeISAtab(isaObject, path = getwd())
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}

\item{path}{A character vector with the name of the directory to which the
file(s) should be written. The default value is the current
working directory.}
}
\description{
Write ISA-Tab object to files. The investigation file, study files and assay
files are written to the folder specified in \code{path}.
}
