% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeStudyFiles}
\alias{writeStudyFiles}
\title{Write study file.}
\usage{
writeStudyFiles(
  isaObject,
  studyFilenames = getStudyInfo(isaObject),
  path = getwd()
)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}

\item{studyFilenames}{A character vector indicating the study files that
should be written.}

\item{path}{A character vector with the name of the directory to which the
file should be written. The default value is the current
working directory.}
}
\description{
Write study file.
}
