% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{aFiles}
\alias{aFiles}
\alias{aFiles<-}
\alias{aFiles,ISA-method}
\alias{aFiles<-,ISA-method}
\title{Get and set aFiles.}
\usage{
aFiles(x)

aFiles(x) <- value

\S4method{aFiles}{ISA}(x)

\S4method{aFiles}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the assay files.
In each data.frame at least the following columns are present:\if{html}{\out{<div class=".">}}\preformatted{}\if{html}{\out{</div>}}}
}
\value{
A list of data.frames containing the assay files.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of aFiles data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of aFiles data.frames.
aFiles(isaObject1)

}
