% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo.R
\name{getAssayFileNames}
\alias{getAssayFileNames}
\title{Retrieve the Assay File Name(s) from an ISA object.}
\usage{
getAssayFileNames(isaObject)
}
\arguments{
\item{isaObject}{An object of the \code{\link{ISA-class}}.}
}
\value{
An character vector containing the Assay File Name(s).
}
\description{
Retrieve from an object of the \code{\link{ISA-class}} the Assay File
Name(s).
}
\seealso{
Other ISAinfo: 
\code{\link{getAssayFileNamesPerSample}()},
\code{\link{getAssayFilesPerStudy}()},
\code{\link{getAssayInfo}()},
\code{\link{getAssayMeasurementTypesPerStudy}()},
\code{\link{getAssayMeasurementTypes}()},
\code{\link{getAssayNameToSample}()},
\code{\link{getAssayNames}()},
\code{\link{getAssayTabs}()},
\code{\link{getAssayTechnologyTypesPerStudy}()},
\code{\link{getAssayTechnologyTypes}()},
\code{\link{getDataFileNamesPerAssay}()},
\code{\link{getFactors}()},
\code{\link{getInvestigationInfo}()},
\code{\link{getRawDataFileToSample}()},
\code{\link{getSampleToAssayName}()},
\code{\link{getSampleToRawDataFile}()},
\code{\link{getSamplesPerAssayFileName}()},
\code{\link{getSamplesPerStudy}()},
\code{\link{getSamples}()},
\code{\link{getStudyInfo}()},
\code{\link{getTreatmentGroups}()},
\code{\link{getTreatments}()}
}
\author{
Maikel Verouden
}
\concept{ISAinfo}
