% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iFileName}
\alias{iFileName}
\alias{iFileName<-}
\alias{iFileName,ISA-method}
\alias{iFileName<-,ISA-method}
\title{Get and set iFileName}
\usage{
iFileName(x)

iFileName(x) <- value

\S4method{iFileName}{ISA}(x)

\S4method{iFileName}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A length-one character vector indicating the name of the
investigation file, a string starting with "i_" and ending in ".txt"}
}
\value{
The file name of the investigation file.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the file name for the investigation file in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract investigation file name.
iFileName(isaObject1)

}
