% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{invest}
\alias{invest}
\alias{invest<-}
\alias{invest,ISA-method}
\alias{invest<-,ISA-method}
\title{Get and set invest.}
\usage{
invest(x)

invest(x) <- value

\S4method{invest}{ISA}(x)

\S4method{invest}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A data.frame containing the investigation information. A
data.frame in which at least the following columns are present:
'Investigation Identifier', 'Investigation Title', 'Investigation Description', 'Investigation Submission Date', 'Investigation Public Release Date'.}
}
\value{
A data.frame containing the investigation information.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the investigation data.frame in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract investigation data.frame.
invest(isaObject1)

}
