% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iPubs}
\alias{iPubs}
\alias{iPubs<-}
\alias{iPubs,ISA-method}
\alias{iPubs<-,ISA-method}
\title{Get and set iPubs.}
\usage{
iPubs(x)

iPubs(x) <- value

\S4method{iPubs}{ISA}(x)

\S4method{iPubs}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A data.frame containing the investigation publications
information. A data.frame in which at least the following columns are
present:
\verb{Investigation PubMed ID}, \verb{Investigation Publication DOI}, \verb{Investigation Publication Author List}, \verb{Investigation Publication Title}, \verb{Investigation Publication Status}, \verb{Investigation Publication Status Term Accession Number}, \verb{Investigation Publication Status Term Source REF}.}
}
\value{
A data.frame containing the investigation publications information.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the iPubs data.frame in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract iPubs data.frame.
iPubs(isaObject1)

}
