% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{oSR}
\alias{oSR}
\alias{oSR<-}
\alias{oSR,ISA-method}
\title{Get and set oSR}
\usage{
oSR(x)

oSR(x) <- value

\S4method{oSR}{ISA}(x)
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A data.frame containing the ontology Source Reference
information. A data.frame with the columns
\verb{Term Source Name}, \verb{Term Source File}, \verb{Term Source Version}, \verb{Term Source Description}.}
}
\value{
A data.frame containing the ontology Source Reference information.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the ontology Source Reference (oSR) data.frame in an object of
\code{\link{ISA-class}}.
}
