% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{path}
\alias{path}
\alias{path<-}
\alias{path,ISA-method}
\alias{path<-,ISA-method}
\title{Get and set the file path for an object of the \code{\link{ISA-class}}.}
\usage{
path(x)

path(x) <- value

\S4method{path}{ISA}(x)

\S4method{path}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of the \code{\link{ISA-class}}.}

\item{value}{A length-one character vector indicating the path to an
accessible directory on the system.}
}
\value{
The path to the folder for the object of the \code{\link{ISA-class}}.

The updated object of the \code{\link{ISA-class}}.
}
\description{
Get and set the file path for an object of the \code{\link{ISA-class}}.
}
