% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isatabr-read.R
\name{readISATab}
\alias{readISATab}
\title{Read an ISA-Tab dataset into an R object.}
\usage{
readISATab(path = getwd(), zipfile = NULL, verbose = FALSE)
}
\arguments{
\item{path}{A character vector with the name of the directory in which the
ISA-Tab files are located (if the parameter zipfile is not provided
or if it is equal to NULL), or the name of the directory where the
zip archive containing the ISA-Tab files is located (if the
parameter zipfile is not NULL). The default value is the current
working directory.}

\item{zipfile}{A character vector with the name of the zip archive containing
ISA-Ttab files themselves (without a directory name in the zip
archive). The default value is NULL (specifying that the ISA-
Tab files have not been archived in one zipped file).}

\item{verbose}{A logical vector indicating to show messages for the different
steps, if TRUE, or not to show them, if FALSE (the default
value).}
}
\value{
An object of the \linkS4class{ISA} class.
}
\description{
Reads an ISA-Tab dataset, given as a zip file or as a set of files in a
specific folder, and builds an object of the \code{\link{ISA-class}}.
}
\examples{
\dontrun{
## Example for mass spectrometry dataset
temp <- tempdir()
datafiles <- c(file.path(system.file("cdf/KO", package = "faahKO"),
                         grep(pattern = "CDF",
                              x = dir(system.file("cdf/KO", package = "faahKO")),
                              ignore.case = TRUE,
                              value = TRUE)),
               file.path(system.file("cdf/WT", package = "faahKO"),
                         grep(pattern = "CDF",
                              x = dir(system.file("cdf/WT", package = "faahKO")),
                              value = TRUE)))
file.copy(datafiles, temp, recursive = TRUE)
isafiles <- file.path(system.file(package = "faahKO"),
                      grep(pattern = "txt",
                           x = dir(system.file(package = "faahKO")),
                           value = TRUE))
file.copy(isafiles, temp, recursive = TRUE)
isaObject1 <- readISATab(path = temp)

## Example of readISATab for a mass spectrometry experiment
isazip <- "faahKO-metadata.zip"
isaObject2 <- readISATab(path = file.path(system.file("extdata",
                                                      package = "Risa")),
                         zipfile = isazip,
                         verbose = TRUE)
}

}
\author{
Maikel Verouden
}
