% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sAssays}
\alias{sAssays}
\alias{sAssays<-}
\alias{sAssays,ISA-method}
\alias{sAssays<-,ISA-method}
\title{Get and set sAssays.}
\usage{
sAssays(x)

sAssays(x) <- value

\S4method{sAssays}{ISA}(x)

\S4method{sAssays}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study assays.
In each data.frame at least the following columns are present:
\verb{Study Assay Measurement Type}, \verb{Study Assay Measurement Type Term Accession Number}, \verb{Study Assay Measurement Type Term Source REF}, \verb{Study Assay Technology Type}, \verb{Study Assay Technology Type Term Accession Number}, \verb{Study Assay Technology Type Term Source REF}, \verb{Study Assay Technology Platform}, \verb{Study Assay File Name}.}
}
\value{
A list of data.frames containing the study assays.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of sAssays data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of sAssays data.frames.
sAssays(isaObject1)

}
