% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sFacts}
\alias{sFacts}
\alias{sFacts<-}
\alias{sFacts,ISA-method}
\alias{sFacts<-,ISA-method}
\title{Get and set sFacts.}
\usage{
sFacts(x)

sFacts(x) <- value

\S4method{sFacts}{ISA}(x)

\S4method{sFacts}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study factors.
In each data.frame at least the following columns are present:
\verb{Study Factor Name}, \verb{Study Factor Type}, \verb{Study Factor Type Term Accession Number}, \verb{Study Factor Type Term Source REF}.}
}
\value{
A list of data.frames containing the study factors.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of sFacts data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of sFacts data.frames.
sFacts(isaObject1)

}
