% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sFiles}
\alias{sFiles}
\alias{sFiles<-}
\alias{sFiles,ISA-method}
\alias{sFiles<-,ISA-method}
\title{Get and set sFiles.}
\usage{
sFiles(x)

sFiles(x) <- value

\S4method{sFiles}{ISA}(x)

\S4method{sFiles}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study files.
In each data.frame at least the following columns are present:
'Source Name', 'Term Source REF', 'Sample Name'.}
}
\value{
A list of data.frames containing the study files.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of sFiles data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of sFiles data.frames.
sFiles(isaObject1)

}
