% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sPubs}
\alias{sPubs}
\alias{sPubs<-}
\alias{sPubs,ISA-method}
\alias{sPubs<-,ISA-method}
\title{Get and set sPubs.}
\usage{
sPubs(x)

sPubs(x) <- value

\S4method{sPubs}{ISA}(x)

\S4method{sPubs}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study publications.
In each data.frame at least the following columns are present:
\verb{Study PubMed ID}, \verb{Study Publication DOI}, \verb{Study Publication Author List}, \verb{Study Publication Title}, \verb{Study Publication Status}, \verb{Study Publication Status Term Accession Number}, \verb{Study Publication Status Term Source REF}.}
}
\value{
A list of data.frames containing the study publications.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of sPubs data.frames in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of sPubs data.frames.
sPubs(isaObject1)

}
