% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{study}
\alias{study}
\alias{study<-}
\alias{study,ISA-method}
\alias{study<-,ISA-method}
\title{Get and set study.}
\usage{
study(x)

study(x) <- value

\S4method{study}{ISA}(x)

\S4method{study}{ISA}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study information.
In each data.frame at least the following columns are present:
\verb{Study Identifier}, \verb{Study Title}, \verb{Study Description}, \verb{Study Submission Date}, \verb{Study Public Release Date}, \verb{Study File Name}.}
}
\value{
A list of data.frames containing the study information.

The updated object of class \linkS4class{ISA}.
}
\description{
Get and set the list of study data.frame in an object of
\linkS4class{ISA}.
}
\examples{
## Read example Atwell data set.
isaObject1 <- readISATab(path = file.path(system.file("extdata/Atwell",
                                          package = "isatabr")))

## Extract list of study data.frames.
study(isaObject1)

}
